huajiyu 发布系统代码

+++ app_Web/src/widgets/action/add-comment-form/add-comment-form-model.ts
@@ -58,7 +58,7 @@ export default class AddCommentModel {
       {
         name: 'objectid',
         prop: 'objectid',
-        dataType: 'INT',
+        dataType: 'BIGINT',
       },
       {
         name: 'comment',
